fix: leaderboard visibility, answer-reveal security, OAuth state, Excel import bugs, i18n AI prompts
- Participant leaderboard: show between questions and after session ends; detect
  early host-end and show appropriate message/icon in AudienceSession
- Answer reveal security: call new close-question endpoint before showing correct
  answer on present screen, preventing late submissions (QuizPresent + session_service_async + quiz.py)
- Present screen sidebar: show leaderboard based on total_participants count, not scores
- Redis audience_state cache: guard against None current_question_status crashing
  _write_audience_state_cache when session ends (session_service_async.py)
- Pydantic: suppress model_ namespace warning in GeminiSettings
- Google OAuth: forward state param from URL to backend callback (was silently
  dropped, breaking CSRF validation added in previous commit)
- Excel import: preserve option text with angle brackets by skipping sanitize_plain;
  fix negative_points truncation by migrating points/negative_points columns to Float
- i18n: translate AI generate-questions promptLabel, promptHint, promptPlaceholder
  for all 10 non-English locales (hi, ta, te, ka, bn, gu, es, fr, de, ru)

+++ backend/persistence/models/quiz.py
@@ -1,7 +1,7 @@
 """
 Quiz feature domain models
 """
-from sqlalchemy import Column, Integer, BigInteger, String, Boolean, Enum as SQLEnum, ForeignKey, Text, JSON, DateTime, UniqueConstraint
+from sqlalchemy import Column, Integer, BigInteger, Float, String, Boolean, Enum as SQLEnum, ForeignKey, Text, JSON, DateTime, UniqueConstraint
 from sqlalchemy.sql import func
 from sqlalchemy.dialects.mysql import DATETIME as MYSQL_DATETIME
 from sqlalchemy.orm import relationship
@@ -170,9 +170,9 @@ class Question(Base, TimestampMixin):
     correct_answer_index = Column(Integer, nullable=True)  # 0-3 for MCQ, null for word_cloud
     question_image_url = Column(String(500), nullable=True)  # Optional image for question text
     option_images = Column(JSON, nullable=True)  # Optional images for MCQ options: {"A": "path", "B": "path", ...}
-    points = Column(Integer, nullable=False, default=1, server_default="1")
+    points = Column(Float, nullable=False, default=1, server_default="1")
     max_time_seconds = Column(Integer, nullable=True)
-    negative_points = Column(Integer, nullable=False, default=0, server_default="0")
+    negative_points = Column(Float, nullable=False, default=0, server_default="0")
     is_required = Column(Boolean, nullable=False, default=False, server_default="0")
     answer_explanation = Column(Text, nullable=True)
